@@ -46,9 +46,9 @@
     "mitt": "^3.0.0",
     "plyr": "^3.7.2",
     "tinycolor2": "^1.4.2",
-    "vue": "^3.2.36",
+    "vue": "^3.2.17",
     "vue-router": "^4.0.14",
-    "vue2": "npm:vue@2.6.12"
+    "vue2": "npm:vue@^2.6.12"
   },
   "files": [
     "dist/*",
@@ -65,12 +65,13 @@
     "@rollup/plugin-commonjs": "^22.0.0",
     "@rollup/plugin-node-resolve": "^13.2.1",
     "@rollup/plugin-replace": "^4.0.0",
-    "@storybook/addon-actions": "^6.5.16",
-    "@storybook/addon-essentials": "^6.5.16",
-    "@storybook/addon-links": "^6.5.16",
-    "@storybook/addons": "^6.5.16",
-    "@storybook/theming": "^6.5.16",
-    "@storybook/vue3": "^6.5.16",
+    "@storybook/addon-actions": "6.4.22",
+    "@storybook/addon-essentials": "6.4.22",
+    "@storybook/addon-links": "6.4.22",
+    "@storybook/addon-postcss": "2.0.0",
+    "@storybook/addons": "6.4.22",
+    "@storybook/theming": "6.4.22",
+    "@storybook/vue3": "6.4.22",
     "@vue/cli-plugin-babel": "^4.5.13",
     "@vue/cli-plugin-typescript": "^4.0.0",
     "@vue/compiler-sfc": "^3.1.0",
@@ -79,10 +80,10 @@
     "clean-css": "^5.2.2",
     "clean-css-cli": "4.3.0",
     "cross-env": "^7.0.3",
-    "eslint": "^8.3.0",
-    "eslint-webpack-plugin": "^3.1.1",
+    "eslint": "^8.30.0",
+    "eslint-webpack-plugin": "^3.2.0",
     "minimist": "^1.2.5",
-    "node-sass": "^4.12.0",
+    "node-sass": "^6.0.1",
     "postcss": "^8.0.0",
     "postcss-import": "12.0.1",
     "postcss-inline-svg": "^4.1.0",
@@ -97,14 +98,14 @@
     "rollup-plugin-terser": "^7.0.2",
     "rollup-plugin-vue": "^6.0.0",
     "rollup-plugin-vue2": "^0.8.1",
-    "sass": "^1.50.0",
-    "sass-loader": "^8.0.0",
+    "sass": "^1.58.0",
+    "sass-loader": "^13.2.0",
     "showdown": "^1.9.1",
     "storybook-vue3-router": "^2.2.1",
-    "vue-loader": "^16.8.3",
+    "vue-loader": "^17.0.0",
     "vue-style-loader": "^4.1.3"
   },
   "engines": {
-    "node": ">=14"
+    "node": ">=14.15.0"
   }
 }
